app/requirements.txt
numpy
PyYAML
youtube-transcript-api
soundfile
kokoro
litellm
Flask
Flask-Cors
python-dotenv
rich
google-generativeai